requirements.txt
Django==1.6
MarkupSafe==0.15
Pygments==1.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-ckeditor==4.4.7
django-disqus==0.4.3
django-suit-redactor==0.0.2
django-toolbelt==0.0.1
gunicorn==19.2.1
markdown2==2.3.0
pycrypto==2.4.1
psycopg2==2.5.1
pylint==1.1.0
requests==2.5.1
static==0.4
static3==0.5.1
urllib3==1.10
wsgiref==0.1.2
